Organized executable commands package
