+++ src/main/java/fr/xephi/authme/command/executable/authme/HelpCommand.java
@@ -1,4 +1,4 @@
-package fr.xephi.authme.command.executable;
+package fr.xephi.authme.command.executable.authme;
 
 import fr.xephi.authme.command.CommandParts;
 import fr.xephi.authme.command.ExecutableCommand;
